[centroid] Move translating_transformer to separate file.
Add ctor taking Geometry to transformer.
This allows to encapsulate the origin-retrieving-logic inside the transformer.
Remove assign_origin() function.
